refactor(design-system): 앱바 Building block을 component 패키지로 분리
 - move: PickleAppBarLayout, PickleAppBarLogo 등 7종 → appbar/component/
 - remove: PickleAppBarV2.kt (슬롯 API 레거시) 삭제
